requirements.txt
streamlit>=1.35
requests
groq
python-dotenv
mistralai
streamlit-audiorecorder
 